requirements.txt
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==2.1.4
blinker==1.3
certifi==2018.8.24
chardet==2.3.0
configobj==5.0.6
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
Flask==1.0.2
future==0.16.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
jsonschema==2.6.0
kiwisolver==1.0.1
line-bot-sdk==1.8.0
MarkupSafe==0.23
mistune==0.8.3
nbconvert==5.4.0
nbformat==4.4.0
oauthlib==1.0.3
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prettytable==0.7.2
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycurl==7.43.0
Pygments==2.2.0
PyJWT==1.3.0
pyparsing==2.2.0
pyserial==3.0.1
pytz==2018.5
PyYAML==3.11
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.9.1
rope==0.11.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
urllib3==1.13.1
wcwidth==0.1.7
Werkzeug==0.14.1
webencodings==0.5.1
widgetsnbextension==3.4.1
